fixed bug with pawnMove
